Don't form dependency rules from #includes which are commented out
This one cost a few thousand pounds in developer time.

Most places that parse out #includes at least check to see whether there's
something (like a comment) on the same line as the #include. Although this
missed the obvious cases such as /* ... */ and #if 0 comments, at least
it was something.

One particular call site did not, however. The upshot was that something
like

    #include "mock_file.h"
    // #include "file.h"

would compile and link both mock_file.o *and* file.o, causing a linker
error.