Allow arrays as include values
This commit is the missing half of #17, "feature/dynamic_include_paths".

Allowing arrays as include path values allows include paths to be
generated at runtime, as the result of some ruby operation.

For example:

    :paths:
      :include:
        - <%= ['some_value', 'some_other_value'] %>
        - <%= `build_system_tool_returning_array_of_paths`.split("\n") %>

This allows include paths to be generated by the build tool, without
upsetting ceedling too much.