use_test_definition option used to compile all sources
Based on TE-HiroakiYamazoe#2 and PR #88 preprocessor define with test name should be added to all source files not only for test file to recompile correctly sources for  the new define. This is the same problem like in #369
Lets put use_test_definitions to be handled like standard behavior for the changed defines.