Add absolute headers filepaths for runner generator
Unity framework gives a TEST_CASE() feature. The problem can exist when we try to add as a test function argument with our internal typedef (ex. typedef unsigned int my_uint32;) defined somewhere in the test file includes or in the test file directly. In current ceedling implementation we cannot directly include test source file into generated runner file. The test file and the runner are compiled separately and linked together, so double definition problem exists. But we can add header files to the generated runner to partially solve this issue.

Because headers generated by ceedling are created in relative way to root directory, we cannot use it directly in the generated runner file. Due to bug https://bugs.ruby-lang.org/issues/10011 there was a problem to create relative paths to the runner and absolute path was used instead of it.