Account for different ordering of files when traversing fs
But this should be done for each Dir.glob call, to ensure stability of
the order of the paths specified by the user.  This is important
notably for include directories, where the order of -I options imposes
what file is included when there are homonyms in different
directories.

cf. commit 558c84939fea8ab1649a6202ac13f781a50d9256