Fixed exception handling config bug
`[:project][:use_exceptions]` was to be set automatically if in use as a cmock plugin. The original code did not reference the flattened hash passed in and no default entry is otherwise created in the code. Therefore, the setting was simply missing and causing builds with CEexception to lack search paths and vendor source files.

These changes fix the hash reference and also take into account the possibility that exceptions are specified in the project settings and may differ from the cmock plugin selection (top-level project settings take precedence).