Broke cyclical dependencies for streaminator
All logging now flowing through it as it should (instead of low-level direct calls to stream_wrapper to get around this). This lays the groundwork for eventual drop-in, proper logging with a logging library. It also sets the stage for more better verbosity handling from the command line in a next commit.