🐛 Setting config for incomplete :unity & :cmock
The order of revised defaults and automatic configuration settings handling left the possiblilty of trying to set a value in Unity and CMock configuration that was incomplete.

A user’s configuration might create a partial configuration block that automatic settings tried to fill out but would encounter nil references dpeending on the accompanying logic.

The fix was to reorder how user config, defaults, and auto configuration values are processed.