Add user id to the README to avoid having object files owned by root
There was some discussion in another issue about how to run Ceedling using docker containers and it was advised to use the root user. I am not certain how it was done as there's no `Dockerfile` in this repo, but we should not promote usage of root user inside the docker container, especially since artifacts created by docker container are used by host user (uploaded, debugging,etc).

Solution is to create a user inside the docker container (lets say developer - command is USER) and use that (I think installation of gem will also need to be done as that user. Then the build directory will not be owned by root (uid 1) but by that user (uid:1000) which in most cases is also an initial user of host system. That way files created within docker container have uid and gid 1000 and not 1 as they do with root and user can delete them without using docker container cleanup (`ceedling clobber` or `ceedling clean`).