Add inline function patterns option for cmock
In default project.yml file we were missing the inline_function_patterns, which is usually crucial for when you have a custom inline function definition or macro. Since that is in CMock documentation and treat_inlines is by default to exclude I made sure I commented out the rest.

While at it I also fixed the typo in treat_externs and treat_inlines few lines above.